clean up + octo + snacks
