+++ lua/plugins/git/advanced-git-search.lua
@@ -1,0 +1,3 @@
+return {
+	{ "aaronhallaert/advanced-git-search.nvim" },
+}
