+++ lua/plugins/git/lazygit.lua
@@ -1,0 +1,17 @@
+return {
+	{
+		-- Docs: https://github.com/kdheepak/lazygit.nvim
+		-- custom configs for lazygit
+		-- https://github.com/jesseduffield/lazygit/blob/master/docs/Config.md
+		-- https://github.com/jesseduffield/lazygit/blob/master/docs/Custom_Command_Keybindings.md
+
+		-- '`' for return
+		-- q for quit
+		-- 'o' when in stash tab to edit config
+		"kdheepak/lazygit.nvim",
+		-- optional for floating window border decoration
+		dependencies = {
+			"nvim-lua/plenary.nvim",
+		},
+	},
+}
