+++ lua/plugins/init.lua
@@ -29,6 +29,7 @@ local plugins = {
 	require("plugins.numb"),
 
 	require("plugins.telescope"),
+	require("plugins.snacks"),
 	require("plugins.spectre"),
 	require("plugins.git"),
